+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.6.10 - Unreleased
+## Version 0.6.11 - Unreleased
+
+
+## Version 0.6.10 - Released 2023-04-04
 
 ### Added
 * Added `nan_policy` argument to `kwarray.Sticher`
@@ -13,6 +16,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 * Allow `None` as `indices` in `kwarray.Sticher.add`
 * Change `nan_behavior` to `nan_policy` in `kwarray.RunningStats`
+* Avoiding global torch imports. This may slow down the ArrayAPI, and we should
+  revisit a refactor after removing 3.6 support.
+* Updated `ub.repr2` to `ub.urepr`.
 
 ### Fixed
 * `robust_normalize` now preserves masked arrays.